feat: adds tab bar view and initial profile view
