+++ ExerciseTracker/Sources/ProfileView.swift
@@ -1,0 +1,11 @@
+import SwiftUI
+
+struct ProfileView: View {
+    var body: some View {
+        Text("My profile")
+    }
+}
+
+#Preview {
+    ProfileView()
+}
